@@ -29,64 +29,67 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/module-federation-runtime": "^12.5.2",
-    "@angular/animations": "^12.2.13",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "^12.2.13",
-    "@angular/compiler": "^12.2.13",
-    "@angular/core": "^12.2.13",
-    "@angular/forms": "^12.2.13",
-    "@angular/platform-browser": "^12.2.13",
-    "@angular/platform-browser-dynamic": "^12.2.13",
-    "@angular/router": "^12.2.13",
-    "@dontcode/core": "^0.4.2",
-    "@dontcode/plugin-common": "^0.4.2",
-    "@dontcode/sandbox": "^0.4.2",
-    "@nrwl/angular": "13.1.4",
-    "broadcast-channel": "^4.5.0",
+    "@angular-architects/module-federation-runtime": "^14.2.3",
+    "@angular/animations": "^13.3.2",
+    "@angular/cdk": "^13.3.2",
+    "@angular/common": "^13.3.2",
+    "@angular/compiler": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@angular/forms": "^13.3.2",
+    "@angular/platform-browser": "^13.3.2",
+    "@angular/platform-browser-dynamic": "^13.3.2",
+    "@angular/router": "^13.3.2",
+    "@dontcode/core": "^0.5.2",
+    "@dontcode/plugin-common": "^0.5.2",
+    "@dontcode/sandbox": "^0.5.2",
+    "dexie": "^3.2.1",
+    "@nrwl/angular": "^14.0.2",
+    "broadcast-channel": "^4.11.0",
     "jsonpath-plus": "^6.0.1",
-    "primeflex": "^3.1.0",
+    "primeflex": "^3.1.3",
     "primeicons": "^5.0.0",
-    "primeng": "^12.2.2",
-    "rxjs": "~6.6.7",
-    "tslib": "^2.3.1",
-    "zone.js": "^0.11.4"
+    "primeng": "^13.3.3",
+    "rxjs": "~7.5.5",
+    "tslib": "^2.4.0",
+    "zone.js": "^0.11.5"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "^12.5.2",
-    "@angular-devkit/build-angular": "^12.2.13",
-    "@angular-eslint/eslint-plugin": "12.6.1",
-    "@angular-eslint/eslint-plugin-template": "12.6.1",
-    "@angular-eslint/template-parser": "12.6.1",
-    "@angular/cli": "^12.2.13",
-    "@angular/compiler-cli": "^12.2.13",
-    "@angular/language-service": "^12.2.13",
-    "@nrwl/cli": "13.2.1",
-    "@nrwl/cypress": "13.2.1",
-    "@nrwl/eslint-plugin-nx": "13.2.1",
-    "@nrwl/jest": "13.2.1",
-    "@nrwl/linter": "13.2.1",
+    "@angular-architects/module-federation": "^14.2.3",
+    "@angular-devkit/build-angular": "^13.3.2",
+    "@angular-eslint/eslint-plugin": "^13.2.1",
+    "@angular-eslint/eslint-plugin-template": "^13.2.1",
+    "@angular-eslint/template-parser": "^13.2.1",
+    "@angular/cli": "^13.3.2",
+    "@angular/compiler-cli": "^13.3.2",
+    "@angular/language-service": "^13.3.2",
+    "@nrwl/cli": "14.0.2",
+    "@nrwl/cypress": "14.0.2",
+    "@nrwl/eslint-plugin-nx": "14.0.2",
+    "@nrwl/jest": "14.0.2",
+    "@nrwl/linter": "14.0.2",
     "@nrwl/tao": "13.2.1",
-    "@nrwl/workspace": "13.2.1",
-    "@types/jest": "^27.0.3",
+    "@nrwl/workspace": "14.0.2",
+    "@types/jest": "^27.4.1",
     "@types/node": "^16.11.9",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "cypress": "^7.7.0",
-    "dotenv": "10.0.0",
-    "eslint": "7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-cypress": "^2.11.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "ng-packagr": "12.2.5",
-    "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.3.2",
-    "ts-jest": "27.0.5",
-    "ts-node": "~10.1.0",
-    "typescript": "^4.3.5"
+    "@typescript-eslint/eslint-plugin": "^5.20.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "cypress": "^9.5.4",
+    "fake-indexeddb": "^3.1.7",
+    "dotenv": "^16.0.0",
+    "eslint": "^8.14.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "jest": "^27.5.1",
+    "jest-preset-angular": "^11.1.1",
+    "ng-packagr": "^13.3.1",
+    "postcss": "^8.4.12",
+    "postcss-import": "^14.1.0",
+    "postcss-preset-env": "^7.4.3",
+    "postcss-url": "^10.1.3",
+    "prettier": "^2.6.2",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
   }
 }
+
