@@ -1,2 +1,1 @@
-import('./bootstrap')
-	.catch(err => console.error(err));
+import('./bootstrap').catch((err) => console.error(err));
