@@ -1,47 +1,3 @@
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
-const mf = require("@angular-architects/module-federation/webpack");
-const path = require("path");
-const share = mf.share;
-
-const sharedMappings = new mf.SharedMappings();
-sharedMappings.register(
-  path.join(__dirname, '../../tsconfig.base.json'),
-  [/* mapped paths to share */]);
-
-module.exports = {
-  output: {
-    uniqueName: "dontCodeSeedPlugin",
-    publicPath: "auto"
-  },
-  optimization: {
-    runtimeChunk: false
-  },
-  resolve: {
-    alias: {
-      ...sharedMappings.getAliases(),
-    }
-  },
-  plugins: [
-    new ModuleFederationPlugin({
-
-      name: "dontCodeSeedPlugin",
-      filename: "remoteEntry.js",
-      exposes: {
-        './SeedField': './libs/seed/src/lib/seed.module.ts',
-        './SeededEntity': './libs/seed/src/lib/seed.module.ts'
-      },
-
-        shared: share({
-          "@angular/core": { singleton: true, strictVersion: true, requiredVersion: 'auto' },
-          "@angular/common": { singleton: true, strictVersion: true, requiredVersion: 'auto' },
-          "@angular/common/http": { singleton: true, strictVersion: true, requiredVersion: 'auto' },
-          "@angular/router": { singleton: true, strictVersion: true, requiredVersion: 'auto' },
-          "@angular/forms": { singleton: true, strictVersion: true, requiredVersion: 'auto' },
-          "@dontcode/core": { singleton: true, strictVersion: true},
-          ...sharedMappings.getDescriptors()
-        })
-
-    }),
-    sharedMappings.getPlugin()
-  ],
-};
+const { withModuleFederation } = require('@nrwl/angular/module-federation');
+const config = require('./module-federation.config');
+module.exports = withModuleFederation(config);
